backend edit test changes

+++ DatabaseTest/views.py
@@ -5,7 +5,7 @@ from rest_framework import status
 from Test.models import Test
 from Database.models import Database
 from .models import DatabaseTest
-
+from django.http import JsonResponse
 
 # Create your views here.
 class AddDBTestView(APIView):
@@ -24,18 +24,31 @@ class AddDBTestView(APIView):
 
 class RemoveDBTestView(APIView):
     def delete(self, request):
-       
-            if "query" in request.data:
+            if "testid" in request.data and "query" in request.data:
                 if DatabaseTest.objects.filter(query=request.data["query"]).exists():
                     DatabaseTest.objects.filter(name=request.data["query"]).delete()
                     return Response(status = status.HTTP_200_OK)
             return Response(status = status.HTTP_400_BAD_REQUEST)
 
+class RemoveDBTestView_byTest_id(APIView):
+    def delete(self, request):
+            if "testid" in request.data:
+                if DatabaseTest.objects.filter(Test_id=request.data["testid"]).exists():
+                    DatabaseTest.objects.filter(Test_id=request.data["testid"]).delete()
+                    return Response(status = status.HTTP_200_OK)
+            return Response(status = status.HTTP_400_BAD_REQUEST)
 class GetAllDBTestsView(APIView):
 
     def get(self, request):
-       
-            if len(DatabaseTest.objects.all()) > 0:
-                DatabaseTest.objects.all()
-                return Response(status = status.HTTP_200_OK)
+        dbtst = list(DatabaseTest.objects.all().values())
+        return JsonResponse(dbtst,safe = False)
+            
+
+class GetDBTestView(APIView):
+    def post(self, request):
+        if "testid" in request.data:
+            if DatabaseTest.objects.filter(Test_id_id = request.data["testid"]).exists():
+                dbtst = list(DatabaseTest.objects.filter(Test_id_id = request.data["testid"]).values())
+                return JsonResponse(dbtst,safe = False)
             return Response(status = status.HTTP_400_BAD_REQUEST)
+        return Response(status = status.HTTP_400_BAD_REQUEST)
